+++ octofludb/classifiers/flucrew.py
@@ -301,7 +301,7 @@ class SegmentToken(Token):
                 and other.typename != self.typename
             ):
                 g.add((uri, P.sameAs, other.as_uri()))
-            elif not other.choose_field_name() in STRAIN_FIELDS:
+            elif not other.choose_field_name() in STRAIN_FIELDS and other.typename is not None:
                 other.object_of(g, uri)